+++ test/support/testValues.js
@@ -1,11 +1,19 @@
 module.exports = {
     host         : "192.168.2.245",
     username     : "08a902b95915cdd9b75547cb50892dc4",
+
     lightsCount  : 16,
+
     locateTimeout: 7000,
     maxScheduleNameLength: 32,
+
     testLightId: 6,
     hueLightId: 6,
     luxLightId: 15,
-    livingColorLightId: 1
+    livingColorLightId: 1,
+
+    version: {
+        api: "1.5.0",
+        software: "01018228"
+    }
 };